+++ Game Files/Frogger.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "../Common Files/Game.h"
+
+class Frogger : public Game {
+public:
+    void OutputsGameLoop();
+};