+++ Game Files/Ghostbusters.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "../Common Files/Game.h"
+
+class Ghostbusters : public Game {
+public:
+    void OutputsGameLoop();
+};